+++ minisat/src/core.rs
@@ -284,7 +284,7 @@ impl SolveStats for Minisat {
     fn max_var(&self) -> Option<Var> {
         let max_var_idx = unsafe { ffi::cminisat_n_vars(self.handle) };
         if max_var_idx > 0 {
-            Some(Var::new((max_var_idx - 1) as usize))
+            Some(Var::new((max_var_idx - 1) as u32))
         } else {
             None
         }
